Made prefix universal
Each command module is now a function. When it is included, it is passed the prefix, and it returns all its properties as it previously did. The difference is these properties can dynamically change based on the prefix before they are returned, allowing their usage docs to include the correct prefix.