Added command folders and config
The program now looks not only in the commands folder for command modules, but also any folder within the commands folder.

A command is not loaded unless there is a JSON file of the same name (which serves as a configuration file for that command).
The command only loads if it has a config file, and if that config file has an 'enabled' property set to true.